test(integration): split fate-live — args-scoped cases shared, global-topic:posts cases dedicated (ADR 0104 step 7, #1027)
Move the args-scoped Term.definitions + auth-only cases to the run-scoped shared stage
(NS-unique slug isolates the args-scoped topic key) and keep the global topic:posts cases
on a dedicated per-file stage, where they can't read another file's interleaved prependNode.

Fixes #1068
